refactor(project): remove storybook preset
