--- config/storybook-preset-carbon/.babelrc.js
@@ -1,10 +1,0 @@
-/**
- * Copyright IBM Corp. 2021, 2021
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-module.exports = {
-  presets: ['@babel/preset-env', '@babel/preset-react'],
-};
