--- config/storybook-preset-carbon/carbon-theme-addon/src/preset.js
@@ -1,14 +1,0 @@
-/**
- * Copyright IBM Corp. 2021, 2021
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-'use strict';
-
-module.exports = {
-  managerEntries(entry = []) {
-    return [...entry, require.resolve('./register')];
-  },
-};
