+++ packages/react/package.json
@@ -80,6 +80,10 @@
     "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-replace": "^2.4.2",
     "@storybook/addon-actions": "^5.3.21",
+    "@storybook/addon-docs": "^5.3.21",
+    "@storybook/addon-knobs": "^5.3.21",
+    "@storybook/addon-links": "^5.3.21",
+    "@storybook/addon-notes": "^5.3.21",
     "@storybook/addon-storysource": "^5.3.21",
     "@storybook/addons": "^5.3.21",
     "@storybook/react": "^5.3.21",
@@ -97,6 +101,7 @@
     "core-js": "^3.6.5",
     "cross-env": "^5.2.0",
     "css-loader": "^3.4.2",
+    "fast-sass-loader": "^1.5.0",
     "gzip-size": "^6.0.0",
     "lcov2badge": "^0.1.0",
     "mini-css-extract-plugin": "^0.9.0",
@@ -116,7 +121,9 @@
     "rollup-plugin-strip-banner": "^2.0.0",
     "rollup-plugin-terser": "^7.0.2",
     "rtlcss": "^2.4.0",
-    "storybook-preset-carbon": "^0.6.0",
+    "sass": "^1.32.2",
+    "sass-loader": "^10.1.1",
+    "storybook-readme": "^5.0.8",
     "string-replace-loader": "^2.1.0",
     "terser-webpack-plugin": "^2.3.2",
     "webpack": "^4.41.5",
